% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMOO.R
\name{scMOO}
\alias{scMOO}
\title{use scMOO to impute dropout values in scRNA-seq data}
\usage{
scMOO(Y.count, W = NULL, K = 0, q = 10, percent = 0.05,
  lambda1 = NULL, lambda2 = NULL, lambda3 = 1e+10, alpha = NULL,
  MAX.ITER = 4, ABSTOL = 0.001, learning.rate = 1e-04,
  epochs = 100, verbose = TRUE, estimates.only = FALSE)
}
\arguments{
\item{W}{W_{gc} is set to the ratio of the number of non-zero elements in Y to the number
of zero elements in Y if Y_{gc} = 0, and to 1 otherwise.}

\item{q}{The parameter for number of additional power iterations.}

\item{percent}{The expression count matrix is preprocessed by filtering out the genes
expressed in at most percent*\eqn{100\%} of the cells.}

\item{lambda1}{Tuning parameter for entropy regularizer.}

\item{lambda2}{Tuning parameter to facilitate feature selection and regularization.}

\item{lambda3}{Tuning parameter to penalize the diagonal elements of the parameter to
eliminate the trivial solution of representing an expression level as a linear combination
of itself.}

\item{alpha}{Tuning parameter to balance the error between the imputed and
observed data and the error of model fitting data.}

\item{ABSTOL}{Absolute tolerance of circulation.}

\item{epochs}{The number of the entire training set going through the entire network.}

\item{verbose}{Whether to output the value of metrics at the end of each epoch. Default is TRUE.}

\item{Y_count}{An expression count matrix. The rows correspond to genes and
the columns correspond to cells. Can be sparse.}

\item{k}{The rank of the low-rank approximation matrix.}

\item{MAX_ITER}{Maximum iteration of scMOO.}

\item{learning_rate}{A hyper-parameter that controls the speed of adjusting the weights of the network
with respect to the loss gradient.}
}
\value{
If `estimates.only = TRUE', then a matrix of scMOO estimates.

If `estimates.only = FALSE', a list with the following components

\item{\code{estimate}}{Recovered (normalized) expression.}

\item{\code{size.factor}}{Size factor used for normalization.}

\item{\code{pred.time}}{Total time for scMOO estimation.}

\item{\code{alpha}}{Tuning parameter to balance the error between the imputed and
observed data and the error of model fitting data.}

\item{\code{w}}{The combination weights W1, W2 and W3.}
}
\description{
use scMOO to impute dropout values in scRNA-seq data
}
\examples{

data("PBMC_CL")

result <- scMOO(PBMC_CL, percent = 0, estimates.only = TRUE)

}
\author{
Ke Jin, \email{kej13@mails.ccnu.edu.cn}
}
